moved all methods from store/mod.rs into state_machine_objects.rs that were using RocksDB reads or writes. Now store/mod.rs is simply an interface to these methods
